Revise training output, add gradient accumulation, add mixed precision,
overfit_to_scene now also fixes the target views, the validation datasets are reset in case of StopIteration.

+++ src/evaluation/metrics.py
@@ -40,8 +40,8 @@ def compute_ssim(
 ) -> Float[Tensor, " batch"]:
     ssim = [
         structural_similarity(
-            gt.detach().cpu().numpy(),
-            hat.detach().cpu().numpy(),
+            gt.detach().cpu().float().numpy(),  # Convert to float in case of bf16 tensors
+            hat.detach().cpu().float().numpy(), # Convert to float in case of bf16 tensors
             win_size=11,
             gaussian_weights=True,
             channel_axis=0,
